+++ CHANGELOG.md
@@ -74,6 +74,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Deactivate hapi fhir client and hibernate logs (configurable)
 - FHIR Terminology Server
 - Location String and Message for validation
+- Parameter LOG_FHIR_VALIDATION
 
 ### Changed
 - Rename Conversion Template as Converter Template